Add rythmic line of poetry
Peotry allows the mind to wander and derive wonders.
Carefully crafted words can captivate an audience and cause a change.
Beaing poetic is not a necessity, but I recommend that you try ocassionally.

This, again, is some random words used solely for the purpose of practice.
Forgive the fact that it has no relationship to development, programming, or tech.
But it does express the power and flavour of the variety of backgrounds in the technical word.

Enjoy your ride reading my logs. Wish me well in my learning journey. I'm late for devotions.